.gitignore creation
